added a scrollview to the db menu and fixed a crash involved in the delete function
the core data stack is pretty loose at the moment so that will need to be cleaned up before going into the next phase

+++ funbun/Model/Coredata.swift
@@ -65,7 +65,9 @@ class Database:ObservableObject {
         do{
             try await context.viewContext.perform {
             let result = try fetchrequest.execute()
-                self.context.viewContext.delete(result[0])
+                if result.count == 0 {} else {self.context.viewContext.delete(result[0])}
+                
+                
             
             
         }
